moved EnhancedWindow.h next to cvui.h, fixed issue with new OpenCV version
