+++ EnhancedWindow.h
