+++ example/src/ui-enhanced-canny/CMakeLists.txt
@@ -4,7 +4,7 @@ include_directories(../../..)
 
 set(ApplicationName ui-enhanced-canny)
 
-add_executable(${ApplicationName} main.cpp ../ui-enhanced-window-component/EnhancedWindow.h)
+add_executable(${ApplicationName} main.cpp ../../../cvui.h ../../../EnhancedWindow.h)
 target_link_libraries( ${ApplicationName} ${OpenCV_LIBS} )
 set_property(TARGET ${ApplicationName} PROPERTY CXX_STANDARD 11)
 install(TARGETS ${ApplicationName} DESTINATION bin)
@@ -15,6 +15,6 @@ if(ADD_PYTHON_EXAMPLES)
         TARGET ${ApplicationName}
 		POST_BUILD
         COMMAND ${CMAKE_COMMAND} -E copy_if_different ${CMAKE_CURRENT_SOURCE_DIR}/*.py ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}
-        COMMAND ${CMAKE_COMMAND} -E copy_if_different ${CMAKE_CURRENT_SOURCE_DIR}/../ui-enhanced-window-component/EnhancedWindow.py ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}
+        COMMAND ${CMAKE_COMMAND} -E copy_if_different ${CMAKE_CURRENT_SOURCE_DIR}/../../../EnhancedWindow.py ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}
 	)
 endif()
